chatbot-service/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-groq
pydantic
python-dotenv


